Add service client class and adjust accordingly
