+++ backstory/src/main/java/com/google/sps/data/perspective/PerspectiveAnalysis.java
@@ -33,21 +33,9 @@ import java.util.Map;
  * piece of text.
  */
 public class PerspectiveAnalysis {
-  /** a private class that keeps the analysis and type for each AttributeType we want analyzed */
-  private class AttributeAnalysis {
-    /** the attribute score given by analysis */
-    private final float score;
-    /** the type that the analysis score is for */
-    private final AttributeType type;
-
-    public AttributeAnalysis(float score, AttributeType type) {
-      this.score = score;
-      this.type = type;
-    }
-  }
 
   /** an array of all the types we want analysis on */
-  private static final AttributeType[] types = {AttributeType.ATTACK_ON_AUTHOR,
+  public static final AttributeType[] ANALYSIS_TYPES = {AttributeType.ATTACK_ON_AUTHOR,
       AttributeType.ATTACK_ON_COMMENTER, AttributeType.FLIRTATION, AttributeType.IDENTITY_ATTACK,
       AttributeType.INCOHERENT, AttributeType.INSULT, AttributeType.LIKELY_TO_REJECT,
       AttributeType.OBSCENE, AttributeType.PROFANITY, AttributeType.SEVERE_TOXICITY,
@@ -60,25 +48,12 @@ public class PerspectiveAnalysis {
   /** the scores & their types from the analysis of the text */
   private final AttributeAnalysis[] analyses;
 
-  public PerspectiveAnalysis(PerspectiveAPI perspective, String text) throws NullPointerException {
+  /** 
+   * Constructs a PerspectiveAnalysis object from text & analyses parameters
+   */ 
+  public PerspectiveAnalysis(String text, AttributeAnalysis[] analyses) {
     this.text = text;
-
-    AnalyzeCommentRequest.Builder builder = new AnalyzeCommentRequest.Builder().comment(
-        new Entry.Builder().type(ContentType.PLAIN_TEXT).text(text).build());
-
-    // add all the types we want to this builder
-    for (AttributeType type : types) {
-      builder.addRequestedAttribute(type, null);
-    }
-
-    AnalyzeCommentRequest request = builder.build();
-    AnalyzeCommentResponse response = perspective.analyze(request);
-
-    analyses = new AttributeAnalysis[types.length];
-
-    for (int i = 0; i < types.length; i++) {
-      analyses[i] = new AttributeAnalysis(fetchScore(response, types[i]), types[i]);
-    }
+    this.analyses = analyses;
   }
 
   /**
@@ -93,16 +68,5 @@ public class PerspectiveAnalysis {
    */
   public AttributeAnalysis[] getAnalyses() {
     return analyses;
-  }
-
-  /**
-   * private helper method to fetch the score for a given response & type
-   *
-   * @return the score for the response for a given type
-   */
-  private float fetchScore(AnalyzeCommentResponse response, AttributeType type) {
-    AttributeScore attributeScore = response.getAttributeScore(type);
-    Score score = attributeScore.getSummaryScore();
-    return score.getValue();
   }
 }
